some refactoring

+++ WeatherService/Model/WeatherData.cs
@@ -9,8 +9,8 @@ namespace WeatherService.Model
         public string StationId { get; set; }
         public string ZipCode { get; set; }
         public DateTime RDate { get; set; }
-        public int? HighTmp { get; set; }
-        public int? LowTmp { get; set; }
+        public double? HighTmp { get; set; }
+        public double? LowTmp { get; set; }
         public double? AvgTmp { get; set; }
         public double? DewPt { get; set; }
     }
